Unnumber VM peering addresses
VirtualBox plays 'badly' with IPv6:
https://www.virtualbox.org/ticket/14855
As a result, configuring IPv6 addresses on the interfaces causes vagrant
to crash randomly (by reporting a link-local address to connect to
instead of a routable address).
As we'll never use these interfaces directly, we specify dummy IPv4
addresses that will be ignored when bridging the interfaces.

Signed-off-by: Olivier Tilmans <olivier.tilmans@uclouvain.be>